python/backend-challenge-flask/pyproject.toml
[tool.poetry]
name = "Resilience FastAPI Challenge"
version = "0.1.0"
description = "Sample subscription app for Resilience backend engineer applicants"
authors = ["Russ Snyder <russ@arceo.ai>"]

[tool.poetry.dependencies]
python = "^3.7"

flask-restplus = "*"
flask-marshmallow = "*"
flask-sqlalchemy = "*"
flask-bcrypt = "*"
marshmallow-sqlalchemy = "*"
sqlalchemy = "*"
werkzeug = "==0.16.1"


[tool.poetry.dev-dependencies]
black = "21.5b2"
Faker = "^8.5.1"
isort = "^5.8.0"
pytest = "^6.2.4"
requests = "^2.25.1"
ipdb = "*"

[tool.black]
line-length = 88
target-version = ['py37']
include = '\.pyi?$'

[tool.isort]
profile = "black"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
